Replace old apply implementation files
